refactor setup-ui into fedimintd
