+++ fedimint-server/Cargo.toml
@@ -43,16 +43,6 @@ tracing ="0.1.37"
 url = { version = "2.3.1", features = ["serde"] }
 threshold_crypto = { git = "https://github.com/jkitman/threshold_crypto", branch = "upgrade-threshold-crypto-libs" }
 
-# setup dependencies
-askama = { version = "0.11.1", features = ["with-axum"] }
-http = "0.2"
-http-body = "0.4"
-askama_axum = "0.1.0"
-axum = { version = "0.4", default-features = false }
-hyper = { version = "0.14", features = ["full"] }
-tower = { version = "0.4", features = ["util"] }
-qrcode-generator = "4.1.6"
-
 [dev-dependencies]
 tracing-subscriber = { version = "0.3.16", features = [ "env-filter" ] }
