+++ fedimint-server/src/lib.rs
@@ -42,9 +42,6 @@ pub mod net;
 /// Fedimint toplevel config
 pub mod config;
 
-/// Admin UI
-pub mod ui;
-
 /// Some abstractions to handle randomness
 mod rng;
